@@ -16,34 +16,33 @@
 		"test:unit": "vitest"
 	},
 	"devDependencies": {
-		"@playwright/test": "^1.28.1",
-		"@sveltejs/adapter-auto": "^3.0.0",
-		"@sveltejs/kit": "^2.0.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.0",
-		"@twa-dev/types": "^7.0.0",
-		"@types/eslint": "8.56.0",
-		"@types/node": "^20.11.8",
-		"@typescript-eslint/eslint-plugin": "^6.0.0",
-		"@typescript-eslint/parser": "^6.0.0",
+		"@playwright/test": "^1.42.1",
+		"@sveltejs/adapter-auto": "^3.1.1",
+		"@sveltejs/kit": "^2.5.4",
+		"@sveltejs/vite-plugin-svelte": "^3.0.2",
+		"@types/eslint": "8.56.6",
+		"@types/node": "^20.11.30",
+		"@typescript-eslint/eslint-plugin": "^7.4.0",
+		"@typescript-eslint/parser": "^7.4.0",
 		"@waku-objects/luminance": "^2.0.1",
 		"dictionary-cs": "^4.0.0",
 		"dictionary-de": "^3.0.0",
 		"dictionary-en": "^4.0.0",
-		"esbuild": "^0.20.0",
+		"esbuild": "^0.20.2",
 		"esbuild-plugin-alias": "^0.2.1",
-		"eslint": "^8.56.0",
+		"eslint": "^8.57.0",
 		"eslint-config-prettier": "^9.1.0",
 		"eslint-plugin-svelte": "^2.36.0-next.4",
-		"prettier": "^3.1.1",
-		"prettier-plugin-svelte": "^3.1.2",
-		"sass": "^1.70.0",
-		"svelte": "^5.0.0-next.1",
-		"svelte-check": "^3.6.0",
+		"prettier": "^3.2.5",
+		"prettier-plugin-svelte": "^3.2.2",
+		"sass": "^1.72.0",
+		"svelte": "5.0.0-next.66",
+		"svelte-check": "^3.6.8",
 		"svelte-i18n": "^4.0.0",
-		"tslib": "^2.4.1",
-		"typescript": "^5.0.0",
-		"vite": "^5.0.3",
-		"vitest": "^1.2.0"
+		"tslib": "^2.6.2",
+		"typescript": "^5.4.3",
+		"vite": "^5.2.6",
+		"vitest": "^1.4.0"
 	},
 	"type": "module"
 }
