@@ -1,20 +1,15 @@
 <script lang="ts">
-	import { _, t } from 'svelte-i18n'
-	import { getContext } from 'svelte'
-	import type { ThemeStore } from '$lib/stores/theme.svelte'
-
-	const themeStore = getContext<ThemeStore>('theme-store')
-
-	$inspect(themeStore.baseColor, themeStore.darkModeSettings)
+	import { _ } from 'svelte-i18n'
+	import Button from '$lib/components/ui/button.svelte'
 
 	const seed = Math.round(Date.now() / 60000)
 </script>
 
 <ul>
-	<li><a href={`/game/${Date.now()}`}>{$_('home.new-game')}</a></li>
-	<li><a href={`/stats`}>{$_('home.stats')}</a></li>
-	<li><a href={`/settings`}>{$_('home.settings')}</a></li>
-	<li><a href={`/about`}>{$_('home.about')}</a></li>
+	<li><Button variant="ghost" href={`/game/${seed}`}>{$_('home.new-game')}</Button></li>
+	<li><Button variant="ghost" href={`/stats`}>{$_('home.stats')}</Button></li>
+	<li><Button variant="ghost" href={`/settings`}>{$_('home.settings')}</Button></li>
+	<li><Button variant="ghost" href={`/about`}>{$_('home.about')}</Button></li>
 </ul>
 
 <style>
@@ -30,22 +25,5 @@
 
 	li {
 		margin: 0;
-	}
-
-	a {
-		color: var(--paragraph-color);
-		font-family: var(--paragraph-font-family);
-		font-size: var(--paragraph-font-size);
-		font-style: var(--paragraph-font-style);
-		font-weight: var(--paragraph-font-weight);
-		line-height: var(--paragraph-line-height);
-		letter-spacing: var(--paragraph-letter-spacing);
-		text-decoration: none;
-		cursor: pointer;
-		padding: 8px;
-		display: block;
-	}
-	a:hover {
-		text-decoration: underline;
 	}
 </style>
